fix: close out -Wsign-compare in libslic3r, 18 -> 0 (118 of 206 total)
libslic3r now has ZERO -Wsign-compare. The last 13 files: GCodeWriter 2,
SupportParameters 2, TriangleMeshSlicer 2, TreeSupport, HFP, GCodeProcessor,
SeamPlacer, WipeTower, LayerRegion, Model, Preset, OrcaTreeSupport,
OrcaTreeSupport3D.

SupportParameters, TreeSupport, HFP, WipeTower, Model and OrcaTreeSupport3D join
MillingPostProcess, Circle and SVG in having no GCC warnings of any kind now.

One of these is an out-of-bounds read, in TriangleMeshSlicer's cut_mesh:

    while ((*storage)[i_end].is_clockwise() && i_end < storage->size())

The operands are the wrong way round, so the element is dereferenced before the
bounds test can short-circuit. The enclosing for loop guarantees i_end < size on
entry, but this while increments i_end and re-tests immediately, so reaching the
last polygon reads one past the end. Now tests i_end < storage->size() first.
The sign-compare warning is what pointed at the line.

Preset.cpp was one site reported eight times - a template comparing an int32_t
index against opt_init->size(), instantiated for eight option types. One cast
cleared all eight. The enclosing loop already used int32_t(opt_cur->size()), so
the index stays signed to match.

SeamPlacer's segment walk used `i < perimeter.end_index - 1` on a size_t
end_index, which underflows to SIZE_MAX on an empty perimeter; it now compares
i + 1 < end_index.

WipeTower had `m_current_tool != -1` where the member is `size_t m_current_tool = 0`,
so the "no tool" sentinel it was written for cannot occur and the clause is always
true. Written as size_t(-1) to be well-typed and behaviour-identical rather than
deleted - whether that member should be signed is a separate question.

The rest are the usual split: forward index loops become size_t; the two
support-enforce-layer tests cast a config int whose declared minimum is 0; the two
GCodeWriter temperature comparisons cast an int16_t cache member that is
initialised to 0 and only ever assigned from the uint32_t parameter.

Verified. Every one of the 13 TUs compiles with -Wsign-compare at 0 and no other
flag's count moved. Runtime against the pre-campaign binary from eae5ede on the
FrankenPrinter profile, 2 runs compared against every before-run: 10 of 11 panel
configs EXACT, all 7 strict gates among them, plus pins and bedcover and both
thinwalls variants. Only handle differs, and it differs from itself run to run.

Note for anyone reading the Error List: it still shows stale -Wsign-compare rows
for Circle.cpp and SVG.cpp. Those were fixed in 0bd7641 and the incremental build
does not recompile them, so the old rows persist. Recompiling either TU directly
reports 0. The per-file compile is the authority, not the list.

Remaining: 88 -Wsign-compare, all in src/slic3r GUI code, plus 19 -Wcatch-value.

+++ src/libslic3r/Format/HFP.cpp
@@ -175,7 +175,7 @@ void HFP::set_custom_gcode_z(Model &model) const {
     std::string extra;
 
     if (!m_filament_set.empty()) {
-        for (int i = 0; i < m_filament_set.size(); i++) {
+        for (size_t i = 0; i < m_filament_set.size(); i++) {
             model.custom_gcode_per_print_z.gcodes.push_back(CustomGCode::Item{m_layer_height * m_slider_values[i],
                                                                               type, extruder, m_filament_set[i].Color,
                                                                               extra});
